Add DateFromToPart
Add the DateFromToPart which stores a date from and to against a content
item.

Update handler to allow indexing of the date values.

Add the part, record, handler, driver and edit view to allow the full
content editing process for this part.

Add a placement.info with entries for the shapes.

Add migration to create the table and set up the part.