Add projection features
Add a new sub feature which depends on projections.

Date from and to will be usable as sorts in projection queries.

Add projection filters based on the Date From and To values.

Allows you to specify only showing content items where the date from/to is
in the future or not specified.